fix for failed tests
